Fixed a search regex problem.
When searching in a specific conversation list, the slash separating the two uid's was being included in the regex for the first uid.  This change means that you can't search in a jabber conference uid, but I don't think someone would ever be the "owner" of one of those anyways.